Pull Beam and Cone out of VizMovingHead.
